FullText: extract search and results
